🔥 对话流密度收尾：删掉聚合后不可达的 file.write / file.edit 卡壳与 permissionBlock 死管线
聚合落地后这两张卡永远走不到：tier() 判它们不是出组档，转录里一律折进活动块
（真 CLI 一轮实测 fileWriteCards / fileEditCards 均为 0），路由表里的两条注册
形同虚设。卡壳删掉，注册表只留出组档。

它们的正文渲染器 hunk-renderer / content-renderer 仍然活着——活动行的就地展开体
直接用它们（spec 决策 13）。原来这两份正文的唯一覆盖挂在被删的卡测试上，所以
先把「长行横向滚动 / 截断条复制完整内容 / 空内容 / diff 头 +N −N」这些用例搬到
renderer 自己的测试上，再删卡：覆盖跟着活着的代码走，不跟着壳走。

RenderItem.permissionBlock 是另一条死管线：配对逻辑往上挂，但没有任何渲染方读它
（RawToolCard 自己从 toolBlock.toolPermission 取）。删字段与赋值，保留把审批标成
已消费的那一步——那个有真实效果（审批不再单独成项）。

顺带清掉随卡壳一起失效的四个文案 key（fileWrite.title/aria/newBadge、fileEdit.aria），
中英同步；排版守卫的扫描名单跟着正文搬。

前端 245 文件 / 2648 项全绿，tsc exit 0，lint 0 error。
