+++ frontend/src/components/agentre/canonical-tool/registry.test.tsx
@@ -21,35 +21,24 @@ describe("CanonicalToolRouter", () => {
     expect(screen.getByTestId("raw-tool-card")).toBeDefined();
   });
 
-  it("routes file.write → FileWriteCard", () => {
-    const block = {
-      type: "tool_use",
-      toolName: "Write",
-      canonical: {
-        kind: "file.write",
-        fileWrite: { path: "/a.ts", content: "x", lines: 1, bytes: 1 },
-      },
-    } as unknown as ChatBlockData;
-    render(<CanonicalToolRouter toolBlock={block} />);
-    expect(screen.getByTestId("file-write-card")).toBeDefined();
-  });
-
-  it("routes file.edit → FileEditCard", () => {
-    const block = {
-      type: "tool_use",
-      toolName: "Edit",
-      canonical: {
-        kind: "file.edit",
-        fileEdit: {
-          files: [
-            { path: "/a.ts", kind: "modified", hunks: [], plus: 1, minus: 0 },
-          ],
-        },
-      },
-    } as unknown as ChatBlockData;
-    render(<CanonicalToolRouter toolBlock={block} />);
-    expect(screen.getByTestId("file-edit-card")).toBeDefined();
-  });
+  // file.write / file.edit 刻意**不注册**：它们一律折进活动块，正文由活动行的
+  // 展开体直接调 content-renderer / hunk-renderer 渲染，卡壳已删。这里守的是
+  // 「别有人顺手把卡壳加回注册表」——真走到路由也只该是通用回落。
+  it.each(["file.write", "file.edit"])(
+    "leaves %s unregistered so it can only fold into an activity block",
+    (kind) => {
+      const block = {
+        type: "tool_use",
+        toolName: "Write",
+        toolInput: { file_path: "/a.ts", content: "x" },
+        canonical: { kind },
+      } as unknown as ChatBlockData;
+      render(<CanonicalToolRouter toolBlock={block} />);
+      expect(screen.getByTestId("raw-tool-card")).toBeDefined();
+      expect(screen.queryByTestId("file-write-card")).toBeNull();
+      expect(screen.queryByTestId("file-edit-card")).toBeNull();
+    },
+  );
 
   it("routes user.ask → UserAskCard", () => {
     const block = {
